requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
bcrypt==5.0.0
bleach==6.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.121.1
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.9.1
kaggle==1.7.4.5
kiwisolver==1.4.9
langchain==1.0.5
langchain-core==1.0.4
langchain-openai==1.0.2
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.42
matplotlib==3.10.7
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.2.6
openai==2.7.2
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.3
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
slack_sdk==3.38.0
sniffio==1.3.1
SQLAlchemy==2.0.44
SQLAlchemy-Utils==0.42.0
stack-data==0.6.3
starlette==0.49.3
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.12.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
webencodings==0.5.1
xxhash==3.6.0
zstandard==0.25.0
